+++ ydb/core/tx/columnshard/engines/scheme/versions/abstract_scheme.h
@@ -49,6 +49,9 @@ public:
 
     virtual std::optional<ui32> GetColumnIdOptional(const std::string& columnName) const = 0;
     virtual int GetFieldIndex(const ui32 columnId) const = 0;
+    bool HasColumnId(const ui32 columnId) const {
+        return GetFieldIndex(columnId) >= 0;
+    }
 
     ui32 GetColumnId(const std::string& columnName) const;
     std::shared_ptr<arrow::Field> GetFieldByIndex(const int index) const;
