+++ ydb/core/tx/data_events/write_data.h
@@ -3,6 +3,7 @@
 
 #include <ydb/core/tx/long_tx_service/public/types.h>
 #include <ydb/core/formats/arrow/arrow_helpers.h>
+#include <ydb/core/formats/arrow/modifier/subset.h>
 #include <ydb/library/accessor/accessor.h>
 
 #include <ydb/library/actors/core/monotonic.h>
@@ -72,11 +73,11 @@ private:
     YDB_READONLY_DEF(IDataContainer::TPtr, Data);
     YDB_READONLY_DEF(std::shared_ptr<arrow::Schema>, PrimaryKeySchema);
     YDB_READONLY_DEF(std::shared_ptr<NOlap::IBlobsWritingAction>, BlobsAction);
-    YDB_ACCESSOR_DEF(std::optional<TSchemaSubset>, SchemaSubset);
+    YDB_ACCESSOR_DEF(std::optional<NArrow::TSchemaSubset>, SchemaSubset);
 public:
     TWriteData(const TWriteMeta& writeMeta, IDataContainer::TPtr data, const std::shared_ptr<arrow::Schema>& primaryKeySchema, const std::shared_ptr<NOlap::IBlobsWritingAction>& blobsAction);
 
-    const TSchemaSubset& GetSchemaSubsetVerified() const {
+    const NArrow::TSchemaSubset& GetSchemaSubsetVerified() const {
         AFL_VERIFY(SchemaSubset);
         return *SchemaSubset;
     }
